requirements.txt
requests>=2.28.0
requests-html==0.10.0
fake-useragent==0.1.11
beautifulsoup4>=4.11.1
lxml>=4.9.0
Pillow>=9.0
tensorflow>=2.9.1